Remove split view controller
Split view controller on iPhone is supported only for iOS8. Also, it just doesn't play nice with the tab bar controller. Removing it also helps us keep the same code working on both platforms vs customizing for one. If we did decide to customize it will be a different view stack anyways.