Fix issue with the overlaid collection view resetting its contentOffset to zero
When pushing to child view controllers, the overlaid collection view
kept resetting its contentOffset to zero. This is the collectionView
trying to do the right thing. We adjust the content offset of the
overlaid view controller based on scrolling the underlying view
controller, even though it doesn't relate to the overlaid collection
view's contentSize. When pushing view controllers it auto-resets to
zero.

Now we're adding empty cells to the overlaid collection view so that the
contentSize matches the max of the detail views, which fixes the issue.